refactor: Rename `lib/plugins/aws/deploy/lib/validateTemplate.js`
