+++ lib/plugins/aws/deploy/index.js
@@ -13,7 +13,7 @@ const waitForChangeSetCreation = require('../lib/wait-for-change-set-creation');
 const createStack = require('./lib/create-stack');
 const cleanupS3Bucket = require('./lib/cleanup-s3-bucket');
 const uploadArtifacts = require('./lib/uploadArtifacts');
-const validateTemplate = require('./lib/validateTemplate');
+const validateTemplate = require('./lib/validate-template');
 const updateStack = require('../lib/update-stack');
 const ensureValidBucketExists = require('./lib/ensure-valid-bucket-exists');
 const path = require('path');
