+++ lib/plugins/aws/deploy/lib/validate-template.js
