+++ test/unit/lib/plugins/aws/deploy/lib/validate-template.test.js
